Delete ahp common
