+++ src/HE.InvestmentLoans.sln
@@ -98,8 +98,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HE.Investments.Common.WWW.T
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HE.Investment.AHP.Domain.Tests", "HE.Investment.AHP.Domain.Tests\HE.Investment.AHP.Domain.Tests.csproj", "{DAD32348-A1E4-42B5-BB19-879CD4942364}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HE.Investment.AHP.Common", "HE.Investment.AHP.Common\HE.Investment.AHP.Common.csproj", "{4A5DC4AB-26AB-4B24-A8C0-6615992C0A6C}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -404,14 +402,6 @@ Global
 		{DAD32348-A1E4-42B5-BB19-879CD4942364}.Release|Any CPU.Build.0 = Release|Any CPU
 		{DAD32348-A1E4-42B5-BB19-879CD4942364}.Release|x64.ActiveCfg = Release|Any CPU
 		{DAD32348-A1E4-42B5-BB19-879CD4942364}.Release|x64.Build.0 = Release|Any CPU
-		{4A5DC4AB-26AB-4B24-A8C0-6615992C0A6C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{4A5DC4AB-26AB-4B24-A8C0-6615992C0A6C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{4A5DC4AB-26AB-4B24-A8C0-6615992C0A6C}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{4A5DC4AB-26AB-4B24-A8C0-6615992C0A6C}.Debug|x64.Build.0 = Debug|Any CPU
-		{4A5DC4AB-26AB-4B24-A8C0-6615992C0A6C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{4A5DC4AB-26AB-4B24-A8C0-6615992C0A6C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{4A5DC4AB-26AB-4B24-A8C0-6615992C0A6C}.Release|x64.ActiveCfg = Release|Any CPU
-		{4A5DC4AB-26AB-4B24-A8C0-6615992C0A6C}.Release|x64.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -457,7 +447,6 @@ Global
 		{933FBB75-C838-4280-A39F-EE0C6B80734C} = {49D82160-DF00-41AA-A4F1-C6A32CF38179}
 		{E19D8DB0-71C7-40A8-A07B-8B86EB5BD351} = {49D82160-DF00-41AA-A4F1-C6A32CF38179}
 		{DAD32348-A1E4-42B5-BB19-879CD4942364} = {0B392B48-7356-4ED7-AC48-C0F576578B81}
-		{4A5DC4AB-26AB-4B24-A8C0-6615992C0A6C} = {0B392B48-7356-4ED7-AC48-C0F576578B81}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {44D30F0B-F008-4D12-95FB-E893E3AFB749}
